:notebook: move decorations out.
