+++ src/vs/workbench/contrib/notebook/test/testNotebookEditor.ts
@@ -222,7 +222,7 @@ export class TestNotebookEditor implements INotebookEditor {
 		throw new Error('Method not implemented.');
 	}
 
-	setCellSelection(cell: CellViewModel, selection: Range): void {
+	setCellEditorSelection(cell: CellViewModel, selection: Range): void {
 		throw new Error('Method not implemented.');
 	}
